+++ R/Rules-class.R
@@ -749,6 +749,68 @@ IncrementsRelative <- function(intervals, increments) {
   )
 }
 
+# IncrementsRelativeDLT ----
+
+## class ----
+
+#' `IncrementsRelativeDLT`
+#'
+#' @description `r lifecycle::badge("stable")`
+#'
+#' [`IncrementsRelativeDLT`] is the class for increments control based on
+#' relative differences in terms of DLTs.
+#'
+#' @slot dlt_intervals (`integer`)\cr an vector with the left bounds of the
+#'   relevant DLT intervals. This parameters specifies the right bounds of the
+#'   intervals. For example, `dlt_intervals  = c(0, 1, 3)` specifies three
+#'   intervals (sets of DLTs: first, 0 DLT; second 1 or 2 DLTs; and the third
+#'   one, at least 3 DLTs. That means, the right bound of the intervals are
+#'   exclusive to the interval and the last interval goes from the last value to
+#'   infinity.
+#' @slot increments (`numeric`)\cr a vector of the same length with the maximum
+#'   allowable relative increments in the `dlt_intervals`.
+#'
+#' @note This considers all DLTs across all cohorts observed so far.
+#'
+#' @seealso [IncrementsRelativeDLTCurrent()] which only considers the DLTs
+#'   in the current cohort.
+#'
+#' @aliases IncrementsRelativeDLT
+#' @export
+#'
+.IncrementsRelativeDLT <- setClass(
+  Class = "IncrementsRelativeDLT",
+  slots = representation(
+    dlt_intervals = "integer",
+    increments = "numeric"
+  ),
+  prototype = prototype(
+    dlt_intervals = c(0L, 1L),
+    increments = c(2, 1)
+  ),
+  contains = "Increments",
+  validity = v_increments_relative_dlt
+)
+
+## constructor ----
+
+#' @rdname IncrementsRelativeDLT-class
+#'
+#' @param dlt_intervals (`numeric`)\cr see slot definition in [`IncrementsRelativeDLT`].
+#' @param increments (`numeric`)\cr see slot definition in [`IncrementsRelativeDLT`].
+#'
+#' @export
+#' @example examples/Rules-class-IncrementsRelativeDLT.R
+#'
+IncrementsRelativeDLT <- function(dlt_intervals, increments) {
+  assert_integerish(dlt_intervals)
+
+  .IncrementsRelativeDLT(
+    dlt_intervals = safeInteger(dlt_intervals),
+    increments = increments
+  )
+}
+
 # nolint end
 
 # IncrementsNumDoseLevels-class ----
@@ -938,73 +1000,6 @@ IncrementsRelativeParts <- function(dltStart,
                              ...)
 }
 
-
-## --------------------------------------------------
-## Increments control based on relative differences in terms of DLTs
-## --------------------------------------------------
-
-##' Increments control based on relative differences in terms of DLTs
-##'
-##' Note that \code{DLTintervals} is to be read as follows. If for example,
-##' we want to specify three intervals: First 0 DLTs, second 1 or 2 DLTs, and
-##' third at least 3 DLTs, then we specify
-##' \code{DLTintervals} to be \code{c(0, 1, 3)}. That means, the right
-##' bound of the intervals are exclusive to the interval -- the vector only
-##' gives the left bounds of the intervals. The last interval goes from 3 to
-##' infinity.
-##'
-##' @note This considers all DLTs across all cohorts observed so far.
-##'
-##' @slot DLTintervals an integer vector with the left bounds of the relevant
-##' DLT intervals
-##' @slot increments a vector of the same length with the maximum allowable
-##' relative increments in the \code{DLTintervals}
-##'
-##' @seealso [IncrementsRelativeDLTCurrent()] which only considers the DLTs
-##' in the current cohort.
-##'
-##' @example examples/Rules-class-IncrementsRelativeDLT.R
-##' @export
-##' @keywords classes
-.IncrementsRelativeDLT <-
-    setClass(Class="IncrementsRelativeDLT",
-             representation(DLTintervals="integer",
-                            increments="numeric"),
-             prototype(DLTintervals=as.integer(c(0, 1)),
-                       increments=c(2, 1)),
-             contains="Increments",
-             validity=
-                 function(object){
-                     o <- Validate()
-
-                     o$check(identical(length(object@increments),
-                                       length(object@DLTintervals)),
-                             "increments must have same length as DLTintervals")
-                     o$check(! is.unsorted(object@DLTintervals, strictly=TRUE),
-                             "DLTintervals has to be sorted and have unique values")
-                     o$check(all(object@DLTintervals >= 0),
-                             "DLTintervals must only contain non-negative integers")
-
-                     o$result()
-         })
-validObject(.IncrementsRelativeDLT())
-
-
-##' Initialization function for "IncrementsRelativeDLT"
-##'
-##' @param DLTintervals see \code{\linkS4class{IncrementsRelativeDLT}}
-##' @param increments see \code{\linkS4class{IncrementsRelativeDLT}}
-##' @return the \code{\linkS4class{IncrementsRelativeDLT}} object
-##'
-##' @export
-##' @keywords methods
-IncrementsRelativeDLT <- function(DLTintervals,
-                                  increments)
-{
-    .IncrementsRelativeDLT(DLTintervals=safeInteger(DLTintervals),
-                           increments=increments)
-}
-
 #' Increment Control based on Relative Differences and Current DLTs
 #'
 #' @description `r lifecycle::badge("experimental")`
@@ -1012,10 +1007,10 @@ IncrementsRelativeDLT <- function(DLTintervals,
 #' The class is based on the number of DLTs observed in the current cohort,
 #' but not cumulatively over all cohorts so far.
 #'
-#' @slot DLTintervals (`integer`)\cr left bounds of the relevant
+#' @slot dlt_intervals (`integer`)\cr left bounds of the relevant
 #'   DLT intervals.
 #' @slot increments (`numeric`)\cr corresponding maximum allowable
-#' relative increments in the `DLTintervals`.
+#' relative increments in the `dlt_intervals`.
 #'
 #' @seealso [IncrementsRelativeDLT()] which considers all DLTs cumulatively
 #' across cohorts and doses.
@@ -1029,14 +1024,14 @@ IncrementsRelativeDLT <- function(DLTintervals,
 
 #' @rdname IncrementsRelativeDLTCurrent-class
 #'
-#' @param DLTintervals see slot description.
+#' @param dlt_intervals see slot description.
 #' @param increments see slot description.
 #'
 #' @export
-IncrementsRelativeDLTCurrent <- function(DLTintervals = c(0, 1),
+IncrementsRelativeDLTCurrent <- function(dlt_intervals = c(0, 1),
                                          increments = c(2, 1))
 {
-  .IncrementsRelativeDLTCurrent(DLTintervals=safeInteger(DLTintervals),
+  .IncrementsRelativeDLTCurrent(dlt_intervals=safeInteger(dlt_intervals),
                                 increments=increments)
 }
 
@@ -1062,7 +1057,7 @@ IncrementsRelativeDLTCurrent <- function(DLTintervals = c(0, 1),
   setClass(Class="IncrementMin",
            representation(IncrementsList="list"),
            prototype(IncrementsList=
-                       list(IncrementsRelativeDLT(DLTintervals=as.integer(c(0, 1)),
+                       list(IncrementsRelativeDLT(dlt_intervals=as.integer(c(0, 1)),
                                                   increments=c(2, 1)),
                             IncrementsRelative(intervals=c(0, 2),
                                                increments=c(2, 1)))),