+++ examples/Rules-class-IncrementsRelativeDLTCurrent.R
@@ -5,6 +5,6 @@
 #      or maximum increasing the dose by 1.22 if 3 or more DLTs were observed
 
 my_increments <- IncrementsRelativeDLTCurrent(
-  DLTintervals = c(0, 1, 3),
+  dlt_intervals = c(0, 1, 3),
   increments = c(1, 0.33, 0.2)
 )