+++ examples/Rules-method-maxDose-IncrementsRelativeDLT.R
@@ -1,3 +1,4 @@
+# nolint start
 
 # Create the data
 data <- Data(x=c(0.1, 0.5, 1.5, 3, 6, 10, 10, 10),
@@ -12,10 +13,11 @@ data <- Data(x=c(0.1, 0.5, 1.5, 3, 6, 10, 10, 10),
 #   - doubling the dose if no DLTs were yet observed
 #   - only increasing the dose by 1.33 if 1 or 2 DLTs were already observed
 #   - only increasing the dose by 1.2 if at least 3 DLTs were already observed
-myIncrements <- IncrementsRelativeDLT(DLTintervals = c(0, 1, 3),
+myIncrements <- IncrementsRelativeDLT(dlt_intervals = c(0, 1, 3),
                                       increments = c(1, 0.33, 0.2))
 
 # Based on the rule above, we then calculate the maximum dose allowed
 nextMaxDose <- maxDose(myIncrements,
                        data=data)
 
+# nolint end