+++ NEWS.md
@@ -1,4 +1,6 @@
 # Version 1.0.9000.9068
+* Renamed slot `DLTintervals` to `dlt_intervals` in `IncrementsRelativeDLT` class
+  and other dependent classes.
 * Renamed slots in `NextBestTD`, `NextBestTDsamples`, `NextBestMaxGain` and
   `NextBestMaxGainSamples` classes.
 * Removed `Model` class.